created grupo structure and screen

+++ composeApp/src/commonMain/kotlin/org/cidadeape/monitoraorcamento/data/model/empenhos/Anexo.kt
@@ -7,5 +7,5 @@ data class Anexo(
     val siglaUnidadeMedida: String,
     val descricaoAnexo: String,
     val qtdeAnexo: Int,
-    val valorUnitarioAnexo: Float
+    val valorUnitarioAnexo: Double
 )
